feat: Request Line을 분리해 저장하는 기능 추가
