+++ web-application-server-gradle/src/main/java/utils/enums/Series.java
